Kiegeszitve valos szamokra

+++ feladatsor/20200213ab.c
@@ -3,14 +3,14 @@
 int main()
 {
     printf("Add meg a teglalap egyik oldalat (cm-ben): ");
-    int a;
-    scanf("%d", &a);
+    float a;
+    scanf("%f", &a);
     printf("Add meg a teglalap masik oldalat (cm-ben): ");
-    int b;
-    scanf("%d", &b);
+    float b;
+    scanf("%f", &b);
 
-    printf("A teglalap kerulete: %d cm\n", 2 * (a + b));
-    printf("A teglalap terulete: %d cm2\n", a * b);
+    printf("A teglalap kerulete: %.2f cm\n", 2 * (a + b));
+    printf("A teglalap terulete: %.2f cm2\n", a * b);
 
     return 0;
 }
